[Refactor] Improve notification service structure
